requirements.txt
fastapi
uvicorn
python-multipart
openai
huggingface_hub
google-genai
elevenlabs
faster-whisper
numpy
scipy
soundfile
noisereduce
librosa
groq
python-dotenv
pyttsx3

# Database
sqlalchemy[asyncio]
asyncpg
alembic
pgvector
httpx
